Split up tests/vm.rs
